+++ project/app/src/gms/java/org/owntracks/android/ui/welcome/WelcomeActivity.kt
@@ -14,6 +14,7 @@ class WelcomeActivity : BaseWelcomeActivity() {
         listOf(
             introFragment,
             connectionSetupFragment,
+            locationPermissionFragment,
             notificationPermissionFragment,
             osRestrictionsFragment,
             playFragment,