+++ project/app/src/main/java/org/owntracks/android/ui/welcome/fragments/LocationPermissionFragment.kt
@@ -1,13 +1,11 @@
 package org.owntracks.android.ui.welcome.fragments
 
-import android.Manifest.permission.POST_NOTIFICATIONS
 import android.content.Context
 import android.os.Build
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.RequiresApi
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -19,7 +17,8 @@ import org.owntracks.android.ui.mixins.LocationPermissionRequester
 import org.owntracks.android.ui.welcome.WelcomeViewModel
 
 @AndroidEntryPoint
-class LocationPermissionFragment @Inject constructor() : WelcomeFragment(),
+class LocationPermissionFragment @Inject constructor() :
+    WelcomeFragment(),
     ActivityResultCallerWithLocationPermissionCallback {
     private lateinit var binding: UiWelcomeLocationPermissionBinding
 
@@ -31,8 +30,7 @@ class LocationPermissionFragment @Inject constructor() : WelcomeFragment(),
 
     private val locationPermissionRequester = LocationPermissionRequester(this)
 
-    override fun shouldBeDisplayed(context: Context): Boolean =
-        Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU && !requirementsChecker.isNotificationsEnabled()
+    override fun shouldBeDisplayed(context: Context): Boolean = !requirementsChecker.hasLocationPermissions()
 
     @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     override fun onCreateView(
@@ -42,32 +40,24 @@ class LocationPermissionFragment @Inject constructor() : WelcomeFragment(),
     ): View {
         binding = UiWelcomeLocationPermissionBinding.inflate(inflater, container, false)
             .apply {
-                uiFragmentWelcomeNotificationPermissionsRequest.setOnClickListener {
+                uiFragmentWelcomeLocationPermissionsRequest.setOnClickListener {
                     requestLocationPermissions()
                 }
             }
         return binding.root
     }
 
-    private val notificationPermissionRequest =
-        registerForActivityResult(ActivityResultContracts.RequestPermission()) {
-            preferences.userDeclinedEnableNotificationPermissions = !it
-            if (it) {
-                binding.uiFragmentWelcomeNotificationPermissionsRequest.visibility = View.INVISIBLE
-                binding.uiFragmentWelcomeNotificationPermissionsMessage.visibility = View.VISIBLE
-            }
-            viewModel.setWelcomeState(WelcomeViewModel.ProgressState.PERMITTED)
-        }
-
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     private fun requestLocationPermissions() {
-        notificationPermissionRequest.launch(POST_NOTIFICATIONS)
+        locationPermissionRequester.requestLocationPermissions(
+            0,
+            requireContext()
+        ) { shouldShowRequestPermissionRationale(it) }
     }
 
     override fun onResume() {
         super.onResume()
         viewModel.setWelcomeState(
-            if (requirementsChecker.isNotificationsEnabled() || preferences.userDeclinedEnableNotificationPermissions) {
+            if (requirementsChecker.hasLocationPermissions() || preferences.userDeclinedEnableLocationPermissions) {
                 WelcomeViewModel.ProgressState.PERMITTED
             } else {
                 WelcomeViewModel.ProgressState.NOT_PERMITTED
@@ -76,10 +66,14 @@ class LocationPermissionFragment @Inject constructor() : WelcomeFragment(),
     }
 
     override fun locationPermissionGranted(code: Int) {
-        TODO("Not yet implemented")
+        preferences.userDeclinedEnableLocationPermissions = false
+        binding.uiFragmentWelcomeLocationPermissionsRequest.visibility = View.INVISIBLE
+        binding.uiFragmentWelcomeLocationPermissionsMessage.visibility = View.VISIBLE
+        viewModel.setWelcomeState(WelcomeViewModel.ProgressState.PERMITTED)
     }
 
     override fun locationPermissionDenied(code: Int) {
-        TODO("Not yet implemented")
+        preferences.userDeclinedEnableLocationPermissions = true
+        viewModel.setWelcomeState(WelcomeViewModel.ProgressState.PERMITTED)
     }
 }