+++ project/app/src/oss/java/org/owntracks/android/ui/welcome/WelcomeActivity.kt
@@ -9,6 +9,7 @@ class WelcomeActivity : BaseWelcomeActivity() {
         listOf(
             introFragment,
             connectionSetupFragment,
+            locationPermissionFragment,
             notificationPermissionFragment,
             osRestrictionsFragment,
             finishFragment